pyproject.toml
[project]
name = "main"
version = "0.1.0"
description = ""
authors = [ {name = "ADMINion",email = "ilgiz_g93@mail.ru"} ]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi (==0.116.1)", 
    "sqlalchemy (==2.0.42)", 
    "alembic (==1.16.4)", 
    "pydantic-settings (>=2.10.1,<3.0.0)", 
    "python-dotenv (>=1.1.1,<2.0.0)", 
    "asyncpg (==0.30.0)", 
    "uvicorn (>=0.35.0,<0.36.0)", 
    "bcrypt (>=4.3.0,<5.0.0)", 
    "pyjwt[crypto] (>=2.10.1,<3.0.0)",
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
